+++ app/src/main/java/com/cogu/spylook/adapters/cards/CuentaCardAdapter.kt
@@ -11,11 +11,13 @@ import android.view.ViewGroup
 import android.widget.ImageView
 import android.widget.PopupWindow
 import android.widget.TextView
+import androidx.appcompat.app.AlertDialog
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import com.cogu.spylook.R
 import com.cogu.spylook.database.AppDatabase
 import com.cogu.spylook.mappers.ContactoToMiniCard
+import com.cogu.spylook.model.cards.ContactoCardItem
 import com.cogu.spylook.model.cards.ContactoMiniCard
 import com.cogu.spylook.model.cards.CuentaCardItem
 import com.cogu.spylook.model.entity.Anotable
@@ -102,20 +104,28 @@ class CuentaCardAdapter(
                 val buttonEliminar = popupView.findViewById<TextView>(R.id.buttonEliminar)
                 buttonEliminar.setOnClickListener {view: View? ->
                     view?.performHapticFeedback(HapticFeedbackConstants.KEYBOARD_TAP)
-                    val dao = AppDatabase.getInstance(context)!!.cuentaDAO()!!
-                    runBlocking {
-                        dao.deleteCuentaAnotable(cardItem.idAnotable)
-                        val index = cardItemList.indexOf(cardItem)
-                        recyclerAnimator.deleteItemWithAnimation(
-                            holder.itemView,
-                            index,
-                            onEmptyCallback = {
-                                cardItemList.add(CuentaCardItem.DEFAULT_FOR_ADD)
-                            },
-                            afterDeleteCallBack = {
-                                popupWindow.dismiss()
-                            })
-                    }
+                    AlertDialog.Builder(context)
+                        .setTitle("Eliminar cuenta")
+                        .setMessage("Desea eliminar por completo la cuenta \"${cardItem.nombre}\"? \n\nEste proceso no se puede deshacer.")
+                        .setPositiveButton("Continuar") { dialog, which ->
+                            val dao = AppDatabase.getInstance(context)!!.cuentaDAO()!!
+                            runBlocking {
+                                dao.deleteCuentaAnotable(cardItem.idAnotable)
+                                val index = cardItemList.indexOf(cardItem)
+                                recyclerAnimator.deleteItemWithAnimation(
+                                    holder.itemView,
+                                    index,
+                                    onEmptyCallback = {
+                                        cardItemList.add(CuentaCardItem.DEFAULT_FOR_ADD)
+                                    },
+                                    afterDeleteCallBack = {
+                                        popupWindow.dismiss()
+                                    })
+                            }
+                        }.setNegativeButton("Cancelar") { dialog, which ->
+                            dialog.dismiss()
+                            popupWindow.dismiss()
+                        }.show()
                 }
                 val x = view!!.getTag(R.id.touch_event_x) as Int
                 val y = view.getTag(R.id.touch_event_y) as Int
